made some feature improvement and fixed bugs
