+++ public/AI-Prompt-Optimizer/js/templates.js
@@ -1,0 +1,96 @@
+export const templates = {
+  coding:
+`Act as a senior software engineer.
+
+Task:
+{PROMPT}
+
+Tone: {TONE}
+Length: {LENGTH}
+
+Requirements:
+- Explain step-by-step
+- Follow best practices
+- Include code examples
+- Mention edge cases
+- Optimize performance where possible
+
+Output format:
+Structured explanation with code.`,
+
+  writing:
+`Act as a professional content writer.
+
+Task:
+{PROMPT}
+
+Tone: {TONE}
+Length: {LENGTH}
+
+Requirements:
+- Clear and engaging tone
+- Well-structured sections
+- Strong introduction
+- Strong conclusion
+- Improve readability
+
+Output format:
+Polished article.`,
+
+  resume:
+`Act as an experienced recruiter.
+
+Task:
+{PROMPT}
+
+Tone: {TONE}
+Length: {LENGTH}
+
+Requirements:
+- ATS-friendly wording
+- Strong action verbs
+- Professional formatting
+- Quantifiable achievements where possible
+- Industry-standard language
+
+Output format:
+Professional resume content.`,
+
+  marketing:
+`Act as a digital marketing strategist.
+
+Task:
+{PROMPT}
+
+Tone: {TONE}
+Length: {LENGTH}
+
+Requirements:
+- Target audience identification
+- Persuasive messaging
+- Strong CTA
+- Marketing-focused tone
+- Conversion optimization
+
+Output format:
+Marketing-ready content.`,
+
+  study:
+`Act as an expert tutor.
+
+Task:
+{PROMPT}
+
+Tone: {TONE}
+Length: {LENGTH}
+
+Requirements:
+- Simple explanations
+- Examples included
+- Key takeaways
+- Step-by-step breakdown
+- Beginner-friendly language
+
+Output format:
+Educational explanation.`
+};
